wip slice setup
