+++ App.js
@@ -1,73 +1,18 @@
 import React from 'react';
-import { NavigationContainer } from '@react-navigation/native';
-import { createStackNavigator } from '@react-navigation/stack';
-import { createBottomTabNavigator } from '@react-navigation/bottom-tabs';
-import { Ionicons } from '@expo/vector-icons/'; // eslint-disable-line import/no-extraneous-dependencies
+import { Provider } from 'react-redux';
+import { configureStore } from '@reduxjs/toolkit';
 
-import LoginScreen from './src/screens/LoginScreen';
-import SummaryScreen from './src/screens/SummaryScreen';
-import CatalogScreen from './src/screens/CatalogScreen';
-import Colors from './src/constants/Colors';
+import RootNavigator from './src/navigation/RootNavigator';
+import counterSlice from './Slice';
 
-const Stack = createStackNavigator();
-const Tab = createBottomTabNavigator();
-
-function selectIconName(name, focused) {
-  if (name === 'Summary') {
-    return focused ? 'md-person-sharp' : 'md-person-outline';
-  } if (name === 'Catalog') {
-    return focused ? 'file-tray-full' : 'file-tray-outline';
-  }
-
-  return '';
-}
-
-// selectScreenOptions is a function that returns a POJO
-const selectScreenOptions = ({ route: { name } }) => ({
-  tabBarIcon: ({ focused, color, size }) => ( // eslint-disable-line react/prop-types
-    <Ionicons
-      name={selectIconName(name, focused)}
-      size={size}
-      color={color}
-    />
-  ),
+const store = configureStore({
+  reducer: counterSlice.reducer,
 });
-
-const PostAuthScreens = () => (
-  <Tab.Navigator
-    screenOptions={selectScreenOptions}
-    tabBarOptions={{
-      activeTintColor: Colors.primary,
-      inactiveTintColor: 'gray',
-    }}
-  >
-    <Tab.Screen name="Summary" component={SummaryScreen} />
-    <Tab.Screen name="Catalog" component={CatalogScreen} />
-  </Tab.Navigator>
-);
 
 export default function App() {
   return (
-    <NavigationContainer>
-      <Stack.Navigator
-        initialRouteName="PreAuth"
-        headerMode="none"
-      >
-        <Stack.Screen
-          name="PreAuth"
-          component={LoginScreen}
-          options={{
-            title: 'Discovery Mobile App',
-          }}
-        />
-        <Stack.Screen
-          name="PostAuth"
-          component={PostAuthScreens}
-          options={{
-            title: 'Discovery Mobile App',
-          }}
-        />
-      </Stack.Navigator>
-    </NavigationContainer>
+    <Provider store={store}>
+      <RootNavigator />
+    </Provider>
   );
 }
